+++ CodexLib/Examples/OnCastDoX.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CodexLib.Examples
+{
+    /// <summary>
+    /// Example: Useful for custom ability logic
+    /// </summary>
+    internal class OnCastDoX : IAbilityOnCastLogic
+    {
+        public void OnCast(AbilityExecutionContext context)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}