+++ DarkCodex/Rogue.cs
@@ -22,10 +22,9 @@ namespace DarkCodex
         [PatchInfo(Severity.Create, "Bleeding Attack", "rogue talent: Bleeding Attack; basic talent: Flensing Strike", false)]
         public static void CreateBleedingAttack()
         {
-            var RogueTalentSelection = Helper.Get<BlueprintFeatureSelection>("c074a5d615200494b8f2a9c845799d93");
             var SneakAttack = Helper.ToRef<BlueprintFeatureReference>("9b9eac6709e1c084cb18c3a366e0ec87");
 
-            BlueprintFeature flensing = Helper.CreateBlueprintFeature(
+            var flensing = Helper.CreateBlueprintFeature(
                 "FlensingStrike",
                 "Flensing Strike",
                 "When you successfully inflict sneak attack damage on a foe with a slashing weapon, your attack doesn’t go particularly deep, but you do carve away a significant portion of skin and flesh. If this sneak attack inflicts bleed damage, the victim of the sneak attack is sickened by the pain and has its natural armor bonus (if any) reduced by a number of points equal to the number of sneak attack dice you possess. These penalties persist as long as the bleed damage persists. Multiple strikes on the same foe do not stack the bleed damage, but the penalty to natural armor does stack, to a maximum penalty equal to the target’s normal full natural armor score."